Move cron functions to their own file
